refactor: split near-threshold large files by responsibility (issue #290)
Split the eight files hovering below the 800-line gate along existing
responsibility seams. Pure file movement; no behavior change.

- gl: platform_x11.go -> platform_state_x11.go, runapp_x11.go, dpi_x11.go
- datagrid: data_source_grid.go -> data_source_grid_request.go,
  data_source_grid_pager.go
- audio: backend_beep.go -> streamers.go, decode.go
- gui: view_listbox.go -> view_listbox_items.go
- gui: inspector.go -> inspector_tree.go, inspector_props.go,
  inspector_strings.go
- gui: view_table.go -> view_table_build.go, view_table_columns.go
- svg: parser.go -> parser_cache.go, parser_animated.go
- gui: window.go -> window_state.go, window_clipboard.go
