+++ gui/audio/backend_beep.go
@@ -3,277 +3,16 @@
 package audio
 
 import (
-	"bytes"
 	"errors"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
-	"sync"
 	"time"
 
 	"github.com/gopxl/beep/v2"
-	"github.com/gopxl/beep/v2/flac"
-	"github.com/gopxl/beep/v2/mp3"
-	"github.com/gopxl/beep/v2/vorbis"
-	"github.com/gopxl/beep/v2/wav"
 )
 
 var _ Backend = (*beepBackend)(nil)
-
-// ---------------------------------------------------------------------------
-// volumeStreamer
-// ---------------------------------------------------------------------------
-
-// volumeStreamer wraps a [beep.Streamer] and multiplies each sample by
-// a volume factor returned by getVolume, clamped to [0, 1].
-type volumeStreamer struct {
-	streamer  beep.Streamer
-	getVolume func() float64
-}
-
-func (v *volumeStreamer) Stream(samples [][2]float64) (n int, ok bool) {
-	n, ok = v.streamer.Stream(samples)
-	vol := clamp01(v.getVolume())
-	if vol < 1 {
-		for i := range n {
-			samples[i][0] *= vol
-			samples[i][1] *= vol
-		}
-	}
-	return
-}
-
-func (v *volumeStreamer) Err() error { return v.streamer.Err() }
-
-// ---------------------------------------------------------------------------
-// fadeStreamer
-// ---------------------------------------------------------------------------
-
-// fadeStreamer wraps a [beep.Streamer] and ramps the volume from
-// startVol to targetVol over a duration.  After the ramp completes:
-//   - for fade-in (targetVol > 0), the streamer becomes transparent.
-//   - for fade-out (targetVol == 0), the streamer drains and calls
-//     onComplete (which must not acquire speaker or channel locks).
-type fadeStreamer struct {
-	streamer   beep.Streamer
-	sampleRate beep.SampleRate
-	startVol   float64
-	targetVol  float64
-	endSamples int
-	elapsed    int
-	done       bool
-	onComplete func()
-}
-
-func (f *fadeStreamer) Stream(samples [][2]float64) (n int, ok bool) {
-	if f.done || f.endSamples <= 0 {
-		if f.targetVol == 0 {
-			if f.onComplete != nil {
-				f.onComplete()
-			}
-			return 0, false
-		}
-		return f.streamer.Stream(samples)
-	}
-	n, ok = f.streamer.Stream(samples)
-	if n == 0 && !ok {
-		return 0, false
-	}
-	for i := range n {
-		t := min(float64(f.elapsed)/float64(f.endSamples), 1)
-		vol := f.startVol + (f.targetVol-f.startVol)*t
-		samples[i][0] *= vol
-		samples[i][1] *= vol
-		f.elapsed++
-	}
-	if f.elapsed >= f.endSamples {
-		f.done = true
-		if f.targetVol == 0 {
-			if f.onComplete != nil {
-				f.onComplete()
-			}
-			return n, false
-		}
-	}
-	return n, ok
-}
-
-func (f *fadeStreamer) Err() error { return f.streamer.Err() }
-
-// ---------------------------------------------------------------------------
-// channelMixer
-// ---------------------------------------------------------------------------
-
-// channelMixer provides N numbered sound-effect channels backed by a
-// single custom [beep.Streamer].  Each channel wraps a [beep.Ctrl] so
-// it can be paused/resumed/replaced without allocations.
-type channelMixer struct {
-	chans        []*beep.Ctrl
-	playing      []bool
-	masterVolume *float64
-	mu           sync.Mutex
-}
-
-func newChannelMixer(n int, master *float64) *channelMixer {
-	chans := make([]*beep.Ctrl, n)
-	for i := range chans {
-		chans[i] = &beep.Ctrl{}
-	}
-	return &channelMixer{
-		chans:        chans,
-		playing:      make([]bool, n),
-		masterVolume: master,
-	}
-}
-
-func (cm *channelMixer) Stream(samples [][2]float64) (n int, ok bool) {
-	var tmp [512][2]float64
-	for len(samples) > 0 {
-		toStream := min(len(tmp), len(samples))
-		clear(samples[:toStream])
-		active := false
-		cm.mu.Lock()
-		for i := range cm.chans {
-			if cm.playing[i] {
-				sn, sok := cm.chans[i].Stream(tmp[:toStream])
-				if sn == 0 && !sok {
-					cm.chans[i].Streamer = nil
-					cm.playing[i] = false
-				} else {
-					for j := range sn {
-						samples[j][0] += tmp[j][0]
-						samples[j][1] += tmp[j][1]
-					}
-					active = true
-				}
-			}
-		}
-		cm.mu.Unlock()
-		if !active {
-			for i := range toStream {
-				samples[i][0] = 0
-				samples[i][1] = 0
-			}
-		} else if *cm.masterVolume < 1 {
-			for i := range toStream {
-				samples[i][0] *= *cm.masterVolume
-				samples[i][1] *= *cm.masterVolume
-			}
-		}
-		samples = samples[toStream:]
-		n += toStream
-	}
-	return n, true
-}
-
-func (cm *channelMixer) Err() error { return nil }
-
-func (cm *channelMixer) firstFree() int {
-	cm.mu.Lock()
-	defer cm.mu.Unlock()
-	for i := range cm.chans {
-		if !cm.playing[i] {
-			return i
-		}
-	}
-	return -1
-}
-
-func (cm *channelMixer) set(ch int, s beep.Streamer) {
-	cm.mu.Lock()
-	cm.chans[ch].Streamer = s
-	cm.chans[ch].Paused = false
-	cm.playing[ch] = true
-	cm.mu.Unlock()
-}
-
-func (cm *channelMixer) halt(channel int) {
-	cm.mu.Lock()
-	defer cm.mu.Unlock()
-	if channel < 0 {
-		for i := range cm.chans {
-			cm.chans[i].Streamer = nil
-			cm.playing[i] = false
-		}
-		return
-	}
-	if channel < len(cm.chans) {
-		cm.chans[channel].Streamer = nil
-		cm.playing[channel] = false
-	}
-}
-
-func (cm *channelMixer) pause(channel int) {
-	cm.mu.Lock()
-	defer cm.mu.Unlock()
-	if channel < 0 {
-		for i := range cm.chans {
-			cm.chans[i].Paused = true
-		}
-		return
-	}
-	if channel < len(cm.chans) {
-		cm.chans[channel].Paused = true
-	}
-}
-
-func (cm *channelMixer) resume(channel int) {
-	cm.mu.Lock()
-	defer cm.mu.Unlock()
-	if channel < 0 {
-		for i := range cm.chans {
-			cm.chans[i].Paused = false
-		}
-		return
-	}
-	if channel < len(cm.chans) {
-		cm.chans[channel].Paused = false
-	}
-}
-
-func (cm *channelMixer) isPlaying(channel int) bool {
-	cm.mu.Lock()
-	defer cm.mu.Unlock()
-	if channel < 0 || channel >= len(cm.chans) {
-		return false
-	}
-	return cm.playing[channel]
-}
-
-func (cm *channelMixer) numChannels() int { return len(cm.chans) }
-
-// ---------------------------------------------------------------------------
-// neverDrain
-// ---------------------------------------------------------------------------
-
-// neverDrain wraps a [beep.Streamer] so it never appears drained.
-// When the inner streamer returns 0,false, the buffer is filled with
-// silence and ok=true is returned.  This prevents the speaker mixer
-// from auto-removing the streamer.
-type neverDrain struct {
-	streamer beep.Streamer
-}
-
-func (n *neverDrain) Stream(samples [][2]float64) (int, bool) {
-	if n.streamer == nil {
-		clear(samples)
-		return len(samples), true
-	}
-	nn, ok := n.streamer.Stream(samples)
-	if !ok {
-		clear(samples[nn:])
-		return len(samples), true
-	}
-	return nn, ok
-}
-
-func (n *neverDrain) Err() error {
-	if n.streamer == nil {
-		return nil
-	}
-	return n.streamer.Err()
-}
 
 // ---------------------------------------------------------------------------
 // musicState
@@ -658,118 +397,4 @@ func (b *beepBackend) ResumeChannel(channel int) {
 
 func (b *beepBackend) IsPlaying(channel int) bool {
 	return b.channels.isPlaying(channel)
-}
-
-// ---------------------------------------------------------------------------
-// decode helpers
-// ---------------------------------------------------------------------------
-
-// decodeReader detects the audio format from file extension and decodes.
-func decodeReader(ext string, rc interface {
-	Read(p []byte) (n int, err error)
-	Seek(offset int64, whence int) (int64, error)
-	Close() error
-}) (beep.StreamSeekCloser, beep.Format, error) {
-	switch ext {
-	case ".wav":
-		stream, format, err := wav.Decode(rc)
-		if err != nil {
-			return nil, beep.Format{}, fmt.Errorf("audio: wav: %w", err)
-		}
-		return stream, format, nil
-	case ".mp3":
-		stream, format, err := mp3.Decode(rc.(io.ReadCloser))
-		if err != nil {
-			return nil, beep.Format{}, fmt.Errorf("audio: mp3: %w", err)
-		}
-		return stream, format, nil
-	case ".ogg":
-		stream, format, err := vorbis.Decode(rc.(io.ReadCloser))
-		if err != nil {
-			return nil, beep.Format{}, fmt.Errorf("audio: ogg: %w", err)
-		}
-		return stream, format, nil
-	case ".flac":
-		stream, format, err := flac.Decode(rc.(io.ReadCloser))
-		if err != nil {
-			return nil, beep.Format{}, fmt.Errorf("audio: flac: %w", err)
-		}
-		return stream, format, nil
-	default:
-		return nil, beep.Format{}, fmt.Errorf("audio: unsupported audio format %q", ext)
-	}
-}
-
-// decodeBytes decodes in-memory audio data.  Detects format by magic
-// bytes and delegates to the appropriate decoder.
-func decodeBytes(data []byte) (beep.StreamSeekCloser, beep.Format, error) {
-	if len(data) < 4 {
-		return nil, beep.Format{}, fmt.Errorf(
-			"audio: data too short (%d bytes)", len(data))
-	}
-
-	switch {
-	case string(data[:4]) == "RIFF":
-		return wav.Decode(newReadSeekCloser(data))
-	case data[0] == 0xFF && len(data) > 1 && data[1]&0xE0 == 0xE0:
-		return mp3.Decode(newReadCloser(data))
-	case string(data[:4]) == "OggS":
-		return vorbis.Decode(newReadCloser(data))
-	case string(data[:4]) == "fLaC":
-		return flac.Decode(newReadCloser(data))
-	default:
-		return nil, beep.Format{}, fmt.Errorf(
-			"audio: unrecognized audio format (magic: % x)", data[:min(4, len(data))])
-	}
-}
-
-// readSeekCloser wraps a [bytes.Reader] into [io.ReadSeekCloser].
-type readSeekCloser struct {
-	r      *bytes.Reader
-	closed bool
-}
-
-func newReadSeekCloser(data []byte) *readSeekCloser {
-	return &readSeekCloser{r: bytes.NewReader(data)}
-}
-
-// readCloser wraps a [bytes.Reader] into [io.ReadCloser].
-type readCloser struct {
-	r      *bytes.Reader
-	closed bool
-}
-
-func newReadCloser(data []byte) *readCloser {
-	return &readCloser{r: bytes.NewReader(data)}
-}
-
-func (r *readCloser) Read(p []byte) (int, error) {
-	if r.closed {
-		return 0, errors.New("audio: reader closed")
-	}
-	return r.r.Read(p)
-}
-
-func (r *readCloser) Close() error {
-	r.closed = true
-	return nil
-}
-
-func (r *readSeekCloser) Read(p []byte) (int, error) {
-	if r.closed {
-		return 0, errors.New("audio: reader closed")
-	}
-	return r.r.Read(p)
-}
-
-func (r *readSeekCloser) Seek(offset int64, whence int) (int64, error) {
-	if r.closed {
-		return 0, errors.New("audio: reader closed")
-	}
-	return r.r.Seek(offset, whence)
-}
-
-func (r *readSeekCloser) Close() error {
-	r.closed = true
-	return nil
 }
