+++ gui/backend/gl/platform_x11.go
@@ -4,15 +4,13 @@ package gl
 
 import (
 	"fmt"
-	"log"
 	"runtime"
-	"strconv"
-	"time"
 
-	gogl "github.com/go-gui-org/go-gui/gui/backend/internal/glbind"
 	"github.com/jezek/xgb"
 	"github.com/jezek/xgb/randr"
 	"github.com/jezek/xgb/xproto"
+
+	gogl "github.com/go-gui-org/go-gui/gui/backend/internal/glbind"
 
 	"github.com/go-gui-org/go-gui/gui"
 	"github.com/go-gui-org/go-gui/gui/backend/ibus"
@@ -38,175 +36,6 @@ const (
 	xcBottomLeftCorner  = 12
 	xcXCursor           = 0
 )
-
-// platformState holds the X11 windowing + EGL state for the GL backend.
-type platformState struct {
-	conn     *xgb.Conn
-	wakeConn *xgb.Conn
-	window   xproto.Window
-
-	eglDpy     uintptr
-	eglConfig  uintptr
-	eglSurface uintptr
-	eglContext uintptr
-
-	cursors   [11]xproto.Cursor
-	curCursor xproto.Cursor
-
-	wmDelete   xproto.Atom
-	wakeAtom   xproto.Atom
-	keymap     *xproto.GetKeyboardMappingReply
-	minKeycode xproto.Keycode
-
-	// Selections. X11 has two independent text buffers: CLIPBOARD (explicit
-	// copy/paste) and PRIMARY (filled by selecting text, pasted with the
-	// middle button). Both are served by the same ownership machinery, so
-	// each keeps its own cached text and owner flag while sharing the atoms
-	// and the read connection below.
-	atomClipboard xproto.Atom
-	atomUTF8      xproto.Atom
-	atomTargets   xproto.Atom
-	atomClipProp  xproto.Atom
-	clipboardText string
-	ownsClipboard bool
-	primaryText   string
-	ownsPrimary   bool
-	clipReadConn  *xgb.Conn     // dedicated connection for reads
-	clipReadWin   xproto.Window // requestor window on clipReadConn
-
-	// Per-monitor DPI (RandR). root anchors monitor queries; curCrtc is
-	// the CRTC the window currently sits on; lastRootXY caches the last
-	// root-relative position so ConfigureNotify only rescans on a move.
-	root        xproto.Window
-	haveRandr   bool
-	curCrtc     randr.Crtc
-	lastRootX   int16
-	lastRootY   int16
-	haveLastPos bool
-
-	physW, physH int32
-	scale        float32
-
-	// Input method. ime is nil when none is reachable, in which case
-	// key presses keep going straight through the keysym path. imeBuf
-	// is reused by drainIME to keep the handoff allocation-free.
-	ime     *ibus.Client
-	imeBuf  []ibus.Event
-	imeEvts []gui.Event
-
-	// Dead-key / Multi_key composition for the raw keysym path (see
-	// compose_x11.go). Only consulted after the input method declines
-	// a key press.
-	compose compose
-
-	w   *gui.Window
-	evt gui.Event // reused per event to avoid per-event allocation
-}
-
-func (p *platformState) makeCurrent() {
-	eglMakeCurrent(p.eglDpy, p.eglSurface, p.eglSurface, p.eglContext)
-}
-
-func (p *platformState) swap() { eglSwapBuffers(p.eglDpy, p.eglSurface) }
-
-func (p *platformState) drawableSize() (int32, int32) { return p.physW, p.physH }
-
-func (p *platformState) dpiScale() float32 {
-	if p.scale <= 0 {
-		return 1
-	}
-	return p.scale
-}
-
-func (p *platformState) setCursor(mc gui.MouseCursor) {
-	if int(mc) >= len(p.cursors) {
-		return
-	}
-	c := p.cursors[mc]
-	if c == 0 {
-		c = p.cursors[gui.CursorDefault]
-	}
-	if c == p.curCursor {
-		return
-	}
-	p.curCursor = c
-	xproto.ChangeWindowAttributes(p.conn, p.window,
-		xproto.CwCursor, []uint32{uint32(c)})
-}
-
-// wake sends a no-op ClientMessage to our own window from a second
-// connection so the event-pump goroutine unblocks from WaitForEvent.
-func (p *platformState) wake() {
-	if p.wakeConn == nil {
-		return
-	}
-	ev := xproto.ClientMessageEvent{
-		Format: 32,
-		Window: p.window,
-		Type:   p.wakeAtom,
-		Data:   xproto.ClientMessageDataUnionData32New([]uint32{0, 0, 0, 0, 0}),
-	}
-	cookie := xproto.SendEventChecked(p.wakeConn, false, p.window, 0, string(ev.Bytes()))
-	_ = cookie.Check() // forces a flush; error is not actionable here
-}
-
-func (p *platformState) destroy() {
-	if p.ime != nil {
-		p.ime.Close()
-		p.ime = nil
-	}
-	if p.conn != nil {
-		for _, c := range p.cursors {
-			if c != 0 {
-				xproto.FreeCursor(p.conn, c)
-			}
-		}
-	}
-	if p.eglDpy != 0 {
-		eglMakeCurrent(p.eglDpy, 0, 0, 0)
-		if p.eglContext != 0 {
-			eglDestroyContext(p.eglDpy, p.eglContext)
-			p.eglContext = 0
-		}
-		if p.eglSurface != 0 {
-			eglDestroySurface(p.eglDpy, p.eglSurface)
-			p.eglSurface = 0
-		}
-		eglTerminate(p.eglDpy)
-		p.eglDpy = 0
-	}
-	if p.conn != nil && p.window != 0 {
-		xproto.DestroyWindow(p.conn, p.window)
-		p.window = 0
-	}
-	if p.wakeConn != nil {
-		p.wakeConn.Close()
-		p.wakeConn = nil
-	}
-	if p.clipReadConn != nil {
-		p.clipReadConn.Close()
-		p.clipReadConn = nil
-	}
-	if p.conn != nil {
-		p.conn.Close()
-		p.conn = nil
-	}
-}
-
-// pumpEvents reads X events on a dedicated goroutine and forwards them
-// on ch. It closes ch when the connection ends so the main loop exits.
-func (p *platformState) pumpEvents(ch chan<- xgb.Event) {
-	for {
-		ev, err := p.conn.WaitForEvent()
-		if ev == nil && err == nil {
-			close(ch) // connection closed
-			return
-		}
-		if ev != nil {
-			ch <- ev
-		}
-	}
-}
 
 // New creates an OpenGL 3.3 backend backed by a native X11 window and an
 // EGL context. Pure Go via xgb + purego — no cgo.
@@ -353,224 +182,6 @@ func New(w *gui.Window) (*Backend, error) {
 	return b, nil
 }
 
-// Destroy releases all backend resources.
-func (b *Backend) Destroy() {
-	b.destroyGLResources()
-	b.plat.destroy()
-}
-
-// Run starts the event loop. Blocks until the window is closed.
-func (b *Backend) Run(w *gui.Window) {
-	defer w.WindowCleanup()
-	b.plat.w = w
-	if w.Config.OnInit != nil {
-		w.Config.OnInit(w)
-	}
-	w.SetWakeMainFn(b.plat.wake)
-
-	events := make(chan xgb.Event, 64)
-	go b.plat.pumpEvents(events)
-
-	// drain processes every currently-queued event without blocking.
-	// Returns false when the connection has closed.
-	drain := func() bool {
-		for {
-			select {
-			case ev, ok := <-events:
-				if !ok {
-					return false
-				}
-				b.handleXEvent(ev)
-			default:
-				b.drainIME()
-				return true
-			}
-		}
-	}
-
-	var rendered bool
-	running := true
-	for running {
-		if !drain() || w.CloseRequested() {
-			break
-		}
-		rendered = w.FrameFn()
-		if rendered {
-			b.renderFrame(w)
-		}
-		b.plat.setCursor(w.MouseCursorState())
-		if !rendered {
-			select {
-			case ev, ok := <-events:
-				if !ok {
-					running = false
-				} else {
-					b.handleXEvent(ev)
-				}
-			case <-time.After(100 * time.Millisecond):
-			}
-		}
-	}
-}
-
-// Run initializes the backend, runs the event loop, and cleans up on
-// exit. Panics on error; call RunE for the error-returning variant.
-func Run(w *gui.Window) {
-	if err := runE(w); err != nil {
-		panic(fmt.Sprintf("gl: %v", err))
-	}
-}
-
-// RunE initializes the backend, runs the event loop, and cleans up on
-// exit. Returns an error instead of panicking.
-func runE(w *gui.Window) error {
-	b, err := New(w)
-	if err != nil {
-		return fmt.Errorf("gl: %w", err)
-	}
-	defer b.Destroy()
-	b.Run(w)
-	return nil
-}
-
-// RunApp starts a multi-window event loop. Panics on error; call
-// RunAppE for the error-returning variant.
-func RunApp(app *gui.App, initialWindows ...*gui.Window) {
-	if err := runAppE(app, initialWindows...); err != nil {
-		panic(fmt.Sprintf("gl: %v", err))
-	}
-}
-
-// taggedEvent carries an X event alongside the backend it belongs to so
-// a single channel can multiplex several windows' event pumps.
-type taggedEvent struct {
-	b      *Backend
-	ev     xgb.Event
-	closed bool
-}
-
-// RunAppE starts a multi-window event loop. Each window is created and
-// registered with app, keyed by its X window id. Blocks until the last
-// window closes.
-//
-//nolint:gocyclo // backend event loop
-func runAppE(app *gui.App, initialWindows ...*gui.Window) error {
-	runtime.LockOSThread()
-
-	backends := make(map[uint32]*Backend) // window XID → backend
-	events := make(chan taggedEvent, 128)
-
-	open := func(w *gui.Window) error {
-		b, err := New(w)
-		if err != nil {
-			return err
-		}
-		b.plat.w = w
-		id := uint32(b.plat.window)
-		backends[id] = b
-		app.Register(id, w)
-		w.SetWakeMainFn(b.plat.wake)
-		if w.Config.OnInit != nil {
-			w.Config.OnInit(w)
-		}
-		go func(bk *Backend) {
-			ch := make(chan xgb.Event, 64)
-			go bk.plat.pumpEvents(ch)
-			for ev := range ch {
-				events <- taggedEvent{b: bk, ev: ev}
-			}
-			events <- taggedEvent{b: bk, closed: true}
-		}(b)
-		return nil
-	}
-
-	for _, w := range initialWindows {
-		if err := open(w); err != nil {
-			for _, b := range backends {
-				b.Destroy()
-			}
-			return fmt.Errorf("gl: create window: %w", err)
-		}
-	}
-
-	closeWindow := func(b *Backend) bool {
-		id := uint32(b.plat.window)
-		if w := app.Window(id); w != nil {
-			w.WindowCleanup()
-		}
-		b.Destroy()
-		delete(backends, id)
-		return app.Unregister(id)
-	}
-
-	var rendered bool
-	for len(backends) > 0 {
-		// Drain queued events + window opens.
-		drained := false
-		for !drained {
-			select {
-			case te := <-events:
-				if te.closed {
-					continue
-				}
-				te.b.handleXEvent(te.ev)
-			case cfg := <-app.PendingOpen():
-				if err := open(gui.NewWindow(cfg)); err != nil {
-					log.Printf("gl: open window: %v", err)
-				}
-			default:
-				for _, b := range backends {
-					b.drainIME()
-				}
-				drained = true
-			}
-		}
-
-		// Handle close requests.
-		for _, b := range backends {
-			w := app.Window(uint32(b.plat.window))
-			if w == nil || !w.CloseRequested() {
-				continue
-			}
-			if closeWindow(b) {
-				return nil // last window closed
-			}
-		}
-		if len(backends) == 0 {
-			return nil
-		}
-
-		// Frame + render each window.
-		rendered = false
-		for _, b := range backends {
-			w := app.Window(uint32(b.plat.window))
-			if w == nil {
-				continue
-			}
-			if w.FrameFn() {
-				b.renderFrame(w)
-				rendered = true
-			}
-			b.plat.setCursor(w.MouseCursorState())
-		}
-
-		if !rendered {
-			select {
-			case te := <-events:
-				if !te.closed {
-					te.b.handleXEvent(te.ev)
-				}
-			case cfg := <-app.PendingOpen():
-				if err := open(gui.NewWindow(cfg)); err != nil {
-					log.Printf("gl: open window: %v", err)
-				}
-			case <-time.After(100 * time.Millisecond):
-			}
-		}
-	}
-	return nil
-}
-
 // --- helpers ---
 
 func internAtom(conn *xgb.Conn, name string) xproto.Atom {
@@ -656,133 +267,4 @@ func loadCursors(p *platformState) {
 	}
 	p.cursors[gui.CursorResizeAll] = load(xcFleur)
 	p.cursors[gui.CursorNotAllowed] = load(xcXCursor)
-}
-
-// readDPIScale derives a UI scale from the Xft.dpi entry in the root
-// RESOURCE_MANAGER property, falling back to 1.0.
-func readDPIScale(conn *xgb.Conn, root xproto.Window) float32 {
-	dpi, err := strconv.Atoi(readXResource(conn, root, "Xft.dpi"))
-	if err == nil && dpi > 0 {
-		return float32(dpi) / 96.0
-	}
-	return 1
-}
-
-// Plausible bounds for a physical display DPI. Values outside this range
-// usually mean a bogus EDID physical size, so the RandR path is rejected
-// in favor of the Xft.dpi fallback.
-const (
-	minPlausibleDPI = 50
-	maxPlausibleDPI = 400
-)
-
-// dpiScaleForWindow computes the UI scale for the monitor containing the
-// root-relative point (x,y), using that monitor's RandR-reported physical
-// size. It falls back to the global Xft.dpi scale when RandR is
-// unavailable or reports no usable physical size. Returns the scale and
-// the CRTC the point lands on (0 when none was resolved).
-func dpiScaleForWindow(conn *xgb.Conn, root xproto.Window, haveRandr bool, x, y int32) (float32, randr.Crtc) {
-	if haveRandr {
-		if s, crtc, ok := randrDPIScale(conn, root, x, y); ok {
-			return s, crtc
-		}
-	}
-	return readDPIScale(conn, root), 0
-}
-
-// randrDPIScale finds the CRTC covering (x,y) and derives a UI scale from
-// its output's physical size. ok is false when RandR data is missing or
-// implausible.
-func randrDPIScale(conn *xgb.Conn, root xproto.Window, x, y int32) (float32, randr.Crtc, bool) {
-	res, err := randr.GetScreenResourcesCurrent(conn, root).Reply()
-	if err != nil || res == nil {
-		return 0, 0, false
-	}
-	for _, crtc := range res.Crtcs {
-		info, ierr := randr.GetCrtcInfo(conn, crtc, res.ConfigTimestamp).Reply()
-		if ierr != nil || info == nil || info.Width == 0 || info.Height == 0 {
-			continue // disabled/disconnected CRTC
-		}
-		if x < int32(info.X) || x >= int32(info.X)+int32(info.Width) ||
-			y < int32(info.Y) || y >= int32(info.Y)+int32(info.Height) {
-			continue
-		}
-		if len(info.Outputs) == 0 {
-			return 0, 0, false
-		}
-		out, oerr := randr.GetOutputInfo(conn, info.Outputs[0], res.ConfigTimestamp).Reply()
-		if oerr != nil || out == nil {
-			return 0, 0, false
-		}
-		if dpi, ok := crtcDPI(info, out); ok {
-			return float32(dpi / 96.0), crtc, true
-		}
-		return 0, 0, false
-	}
-	return 0, 0, false
-}
-
-// crtcDPI averages the horizontal and vertical DPI from the CRTC pixel
-// size and the output's physical millimetre size. ok is false when no
-// physical dimension is reported or the result is implausible.
-func crtcDPI(info *randr.GetCrtcInfoReply, out *randr.GetOutputInfoReply) (float64, bool) {
-	const mmPerInch = 25.4
-	// A 90/270° rotation swaps the pixel axes relative to physical size.
-	pw, ph := float64(info.Width), float64(info.Height)
-	if info.Rotation&(randr.RotationRotate90|randr.RotationRotate270) != 0 {
-		pw, ph = ph, pw
-	}
-	var sum float64
-	var n int
-	if out.MmWidth > 0 {
-		sum += pw / (float64(out.MmWidth) / mmPerInch)
-		n++
-	}
-	if out.MmHeight > 0 {
-		sum += ph / (float64(out.MmHeight) / mmPerInch)
-		n++
-	}
-	if n == 0 {
-		return 0, false
-	}
-	dpi := sum / float64(n)
-	if dpi < minPlausibleDPI || dpi > maxPlausibleDPI {
-		return 0, false
-	}
-	return dpi, true
-}
-
-// maybeRescaleDPI re-evaluates the per-monitor scale when the window has
-// moved to a CRTC with a different DPI, updating plat.scale and the glyph
-// backend. It reports whether the scale changed so the caller can trigger
-// a relayout. ConfigureNotify coordinates are frame-relative under a
-// reparenting WM, so the true root position is queried explicitly and a
-// RandR rescan runs only when that position changed.
-func (b *Backend) maybeRescaleDPI() bool {
-	if !b.plat.haveRandr {
-		return false
-	}
-	t, err := xproto.TranslateCoordinates(b.plat.conn, b.plat.window,
-		b.plat.root, 0, 0).Reply()
-	if err != nil || t == nil {
-		return false
-	}
-	if b.plat.haveLastPos && t.DstX == b.plat.lastRootX &&
-		t.DstY == b.plat.lastRootY {
-		return false // no move → still on the same monitor
-	}
-	b.plat.lastRootX, b.plat.lastRootY = t.DstX, t.DstY
-	b.plat.haveLastPos = true
-
-	scale, crtc := dpiScaleForWindow(b.plat.conn, b.plat.root, true,
-		int32(t.DstX), int32(t.DstY))
-	b.plat.curCrtc = crtc
-	if scale == b.plat.scale {
-		return false
-	}
-	b.plat.scale = scale
-	if b.glyphBack != nil {
-		b.glyphBack.dpiScale = scale
-	}
-	return true
 }
