+++ gui/inspector.go
@@ -2,9 +2,6 @@ package gui
 
 import (
 	"fmt"
-	"slices"
-	"strconv"
-	"strings"
 )
 
 const (
@@ -41,43 +38,6 @@ const (
 	inspectorPropEventsID   = "__prop_events"
 	inspectorPropChildrenID = "__prop_children"
 )
-
-func inspectorNodeTextStyle() TextStyle {
-	return guiTheme.treeStyle.TextStyle
-}
-
-func inspectorNodeIconStyle() TextStyle {
-	return guiTheme.treeStyle.textStyleIcon
-}
-
-type inspectorStackFrame struct {
-	nodes []TreeNodeCfg
-	pos   int
-}
-
-type inspectorNodeProps struct {
-	TypeName    string
-	ID          string
-	events      string
-	TextPreview string
-	Children    int
-	Padding     Padding
-	X           float32
-	Y           float32
-	Width       float32
-	Height      float32
-	Spacing     float32
-	Radius      float32
-	Focusable   bool
-	Scrollable  bool
-	Opacity     float32
-	Color       Color
-	Sizing      Sizing
-	HAlign      HorizontalAlign
-	VAlign      verticalAlign
-	IsFloat     bool
-	Clip        bool
-}
 
 func inspectorToggle(w *Window) {
 	if w == nil {
@@ -204,383 +164,6 @@ func inspectorHelpBar() View {
 	})
 }
 
-func inspectorTreeView(w *Window) View {
-	var nodes []TreeNodeCfg
-	if w != nil {
-		nodes = w.inspectorTreeCache
-	}
-	return Tree(TreeCfg{
-		ID:       inspectorTreeID,
-		indent:   16,
-		Spacing:  SomeF(1),
-		Nodes:    nodes,
-		OnSelect: func(id string, ctx EventCtx) { inspectorSelect(id, ctx.Window) },
-	})
-}
-
-func inspectorSelect(path string, w *Window) {
-	if w == nil {
-		return
-	}
-	if strings.HasPrefix(path, inspectorPropPrefix) {
-		if selected := inspectorSelectedPath(w); selected != "" {
-			treeFocusedSet(w, inspectorTreeID, selected)
-		}
-		return
-	}
-	sm := StateMap[string, string](w, nsInspector, capInspector)
-	// Default "": absent key means nothing is currently selected.
-	selected := sm.GetOr("selected", "")
-	if selected == path {
-		sm.Delete("selected")
-		sm.Delete("scroll_to")
-		treeFocusedSet(w, inspectorTreeID, "")
-		w.UpdateWindow()
-		return
-	}
-
-	sm.Set("selected", path)
-	sm.Set("scroll_to", path)
-	treeFocusedSet(w, inspectorTreeID, path)
-
-	expanded := StateReadOr[string, map[string]bool](
-		w, nsTreeExpanded, inspectorTreeID, nil)
-	if expanded == nil {
-		expanded = make(map[string]bool)
-	}
-	// Mark every ancestor prefix expanded so the tree widget
-	// flattens the path down to the selected node. Paths are
-	// colon-composed (ScopeIDN), so split on IDSep.
-	parts := strings.Split(path, IDSep)
-	prefix := parts[0]
-	expanded[prefix] = true
-	for i := 1; i < len(parts); i++ {
-		prefix = ScopeID(prefix, parts[i])
-		expanded[prefix] = true
-	}
-	StateMap[string, map[string]bool](w, nsTreeExpanded, capModerate).
-		Set(inspectorTreeID, expanded)
-	w.UpdateWindow()
-}
-
-func inspectorPickPath(layout *Layout, x, y float32) string {
-	if layout == nil || len(layout.Children) == 0 {
-		return ""
-	}
-	return inspectorPickRecurse(&layout.Children[0], "0", x, y)
-}
-
-func inspectorPickRecurse(layout *Layout, path string, x, y float32) string {
-	if layout == nil || layout.Shape == nil {
-		return ""
-	}
-	if !layout.Shape.PointInShape(x, y) {
-		return ""
-	}
-	for i := range slices.Backward(layout.Children) {
-		childPath := ScopeIDN(path, "", i)
-		if picked := inspectorPickRecurse(
-			&layout.Children[i], childPath, x, y); picked != "" {
-			return picked
-		}
-	}
-	return path
-}
-
-func inspectorSelectedPath(w *Window) string {
-	if w == nil {
-		return ""
-	}
-	return StateReadOr(w, nsInspector, "selected", "")
-}
-
-func inspectorBuildTreeNodes(
-	w *Window,
-	layout *Layout,
-	selected string,
-	props map[string]inspectorNodeProps,
-) []TreeNodeCfg {
-	if layout == nil || len(layout.Children) == 0 {
-		return nil
-	}
-	var expanded map[string]bool
-	if w != nil {
-		expanded = treeExpandedState(w, inspectorTreeID)
-	}
-	return inspectorLayoutToTree(
-		w, expanded, &layout.Children[0], "0", selected, props)
-}
-
-func inspectorLayoutToTree(
-	w *Window,
-	expanded map[string]bool,
-	layout *Layout,
-	path string,
-	selected string,
-	props map[string]inspectorNodeProps,
-) []TreeNodeCfg {
-	if layout == nil {
-		return nil
-	}
-	propSnapshot := inspectorSnapshotProps(layout)
-	if props != nil {
-		props[path] = propSnapshot
-	}
-
-	childNodes := make([]TreeNodeCfg, 0, 16)
-	if path == selected {
-		childNodes = append(childNodes, inspectorPropsNodes(propSnapshot)...)
-	}
-
-	isExpanded := expanded != nil && expanded[path]
-	isAncestor := !isExpanded && selected != "" && strings.HasPrefix(selected, path+IDSep)
-
-	if isExpanded || isAncestor {
-		for i := range layout.Children {
-			childPath := ScopeIDN(path, "", i)
-			childNodes = append(
-				childNodes,
-				inspectorLayoutToTree(
-					w, expanded, &layout.Children[i], childPath, selected, props,
-				)...,
-			)
-		}
-	} else if len(layout.Children) > 0 {
-		// Add a dummy child to show the arrow icon for collapsed nodes.
-		childNodes = append(childNodes, TreeNodeCfg{
-			ID: ScopeID(path, "__dummy__"),
-		})
-	}
-
-	return []TreeNodeCfg{{
-		ID:            path,
-		Text:          inspectorNodeLabel(layout.Shape),
-		TextStyle:     inspectorNodeTextStyle(),
-		textStyleIcon: inspectorNodeIconStyle(),
-		Nodes:         childNodes,
-	}}
-}
-
-func inspectorPropNode(
-	id, text string, ps, pis TextStyle,
-) TreeNodeCfg {
-	return TreeNodeCfg{
-		ID: id, Text: text,
-		TextStyle: ps, textStyleIcon: pis,
-	}
-}
-
-func inspectorPropsNodes(p inspectorNodeProps) []TreeNodeCfg {
-	propColor := guiTheme.inspectorStyle.colorTextProp
-	ps := TextStyle{
-		Size:  guiTheme.sizeTextXSmall,
-		Color: propColor,
-	}
-	// Icon5 is the XSmall themed icon style, so it already carries the
-	// theme's icon family (app-overridable via ThemeCfg.IconFontFamily).
-	pis := guiTheme.icon5
-	pis.Color = propColor
-
-	nodes := make([]TreeNodeCfg, 0, 16)
-	if p.TextPreview != "" {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropTextID,
-				`text: "`+p.TextPreview+`"`, ps, pis))
-	}
-	if p.ID != "" {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropIDID,
-				"id: "+p.ID, ps, pis))
-	}
-	nodes = append(nodes,
-		inspectorPropNode(inspectorPropPosID,
-			"pos: "+strconv.Itoa(int(p.X))+", "+
-				strconv.Itoa(int(p.Y)), ps, pis),
-		inspectorPropNode(inspectorPropSizeID,
-			"size: "+strconv.Itoa(int(p.Width))+" x "+
-				strconv.Itoa(int(p.Height)), ps, pis),
-	)
-	if p.Sizing.Width != sizingFit || p.Sizing.Height != sizingFit {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropSizingID,
-				"sizing: "+inspectorSizingString(p.Sizing), ps, pis))
-	}
-	if !p.Padding.isNone() {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropPaddingID,
-				"pad: "+strconv.Itoa(int(p.Padding.Top))+" "+
-					strconv.Itoa(int(p.Padding.Right))+" "+
-					strconv.Itoa(int(p.Padding.Bottom))+" "+
-					strconv.Itoa(int(p.Padding.Left)), ps, pis))
-	}
-	if p.Spacing > 0 {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropSpacingID,
-				"spacing: "+strconv.Itoa(int(p.Spacing)), ps, pis))
-	}
-	if p.Color.IsSet() && p.Color.A > 0 {
-		nodes = append(nodes, TreeNodeCfg{
-			ID:        inspectorPropColorID,
-			Text:      "color: " + inspectorColorString(p.Color),
-			Icon:      "\u25A0",
-			TextStyle: ps,
-			textStyleIcon: TextStyle{
-				Size:  guiTheme.sizeTextXSmall,
-				Color: p.Color,
-			},
-		})
-	}
-	if p.Radius > 0 {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropRadiusID,
-				"radius: "+strconv.Itoa(int(p.Radius)), ps, pis))
-	}
-	if p.Focusable {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropFocusID,
-				"focusable: true", ps, pis))
-	}
-	if p.Scrollable {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropScrollID,
-				"scrollable: true", ps, pis))
-	}
-	if p.HAlign != HAlignStart || p.VAlign != VAlignTop {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropAlignID,
-				"align: "+inspectorAlignString(p.HAlign, p.VAlign),
-				ps, pis))
-	}
-	if p.IsFloat {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropFloatID,
-				"float: true", ps, pis))
-	}
-	if p.Clip {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropClipID,
-				"clip: true", ps, pis))
-	}
-	if p.Opacity < 1 {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropOpacityID,
-				fmt.Sprintf("opacity: %.2f", p.Opacity), ps, pis))
-	}
-	if p.events != "" {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropEventsID,
-				"events: "+p.events, ps, pis))
-	}
-	if p.Children > 0 {
-		nodes = append(nodes,
-			inspectorPropNode(inspectorPropChildrenID,
-				"children: "+strconv.Itoa(p.Children), ps, pis))
-	}
-	return nodes
-}
-
-func inspectorSnapshotProps(layout *Layout) inspectorNodeProps {
-	if layout == nil || layout.Shape == nil {
-		return inspectorNodeProps{}
-	}
-	shape := layout.Shape
-	textPreview := ""
-	if shape.TC != nil && shape.TC.Text != "" {
-		textPreview = truncatePreview(shape.TC.Text, 30)
-	}
-
-	props := inspectorNodeProps{
-		TypeName: inspectorTypeName(shape),
-		// The effective ID, not the leaf: it is the string SetFocus,
-		// FindByID and the test helpers take, so it is the one worth
-		// copying out of the inspector.
-		ID:          shape.idKey(),
-		X:           shape.X,
-		Y:           shape.Y,
-		Width:       shape.Width,
-		Height:      shape.Height,
-		Sizing:      shape.Sizing,
-		Padding:     shape.Padding,
-		Spacing:     shape.Spacing,
-		Color:       shape.Color,
-		Radius:      shape.Radius,
-		Focusable:   shape.Focusable,
-		Scrollable:  shape.Scrollable,
-		HAlign:      shape.HAlign,
-		VAlign:      shape.VAlign,
-		IsFloat:     shape.Float,
-		Clip:        shape.Clip,
-		Opacity:     shape.Opacity,
-		TextPreview: textPreview,
-		Children:    len(layout.Children),
-	}
-	if shape.hasEvents() {
-		props.events = inspectorEventsString(shape.events)
-	}
-	return props
-}
-
-func inspectorNodeLabel(shape *Shape) string {
-	if shape == nil {
-		return "(nil)"
-	}
-	label := inspectorTypeName(shape) + " " +
-		strconv.Itoa(int(shape.Width)) + "x" +
-		strconv.Itoa(int(shape.Height))
-	if id := shape.idKey(); id != "" {
-		label += " #" + id
-	}
-	return label
-}
-
-func inspectorTypeName(shape *Shape) string {
-	if shape == nil {
-		return "(nil)"
-	}
-	switch shape.shapeType {
-	case shapeText:
-		return "text"
-	case shapeImage:
-		return "image"
-	case shapeCircle:
-		return "circle"
-	case shapeRTF:
-		return "rtf"
-	case shapeSVG:
-		return "svg"
-	case shapeDrawCanvas:
-		return "draw_canvas"
-	case shapeTermGrid:
-		return "termgrid"
-	case shapeNone, shapeRectangle:
-		switch shape.Axis {
-		case axisTopToBottom:
-			return "column"
-		case axisLeftToRight:
-			return "row"
-		default:
-			return "canvas"
-		}
-	default:
-		return "unknown"
-	}
-}
-
-func inspectorFindByPath(layout *Layout, path string) (*Layout, bool) {
-	if layout == nil || path == "" {
-		return nil, false
-	}
-	node := layout
-	for part := range strings.SplitSeq(path, IDSep) {
-		idx, err := strconv.Atoi(part)
-		if err != nil || idx < 0 || idx >= len(node.Children) {
-			return nil, false
-		}
-		node = &node.Children[idx]
-	}
-	return node, true
-}
-
 func inspectorInjectWireframe(w *Window) {
 	if w == nil {
 		return
@@ -617,44 +200,6 @@ func inspectorInjectWireframe(w *Window) {
 		Color:     guiTheme.inspectorStyle.colorPadding,
 		Thickness: 1,
 	}, w)
-}
-
-func inspectorEventsString(events *eventHandlers) string {
-	if events == nil {
-		return ""
-	}
-	names := make([]string, 0, 10)
-	if events.OnClick != nil {
-		names = append(names, "click")
-	}
-	if events.OnChar != nil {
-		names = append(names, "char")
-	}
-	if events.OnKeyDown != nil {
-		names = append(names, "keydown")
-	}
-	if events.OnMouseMove != nil {
-		names = append(names, "mouse_move")
-	}
-	if events.OnMouseUp != nil {
-		names = append(names, "mouse_up")
-	}
-	if events.OnMouseScroll != nil {
-		names = append(names, "scroll")
-	}
-	if events.onScroll != nil {
-		names = append(names, "scroll_cb")
-	}
-	if events.OnHover != nil {
-		names = append(names, "hover")
-	}
-	if events.onIMECommit != nil {
-		names = append(names, "ime")
-	}
-	if events.AmendLayout != nil {
-		names = append(names, "amend")
-	}
-	return strings.Join(names, ", ")
 }
 
 func inspectorColorString(c Color) string {
@@ -717,50 +262,4 @@ func inspectorFlatRowIndex(
 		}
 	}
 	return -1
-}
-
-func inspectorSizingString(s Sizing) string {
-	return inspectorSizingTypeString(s.Width) + ", " +
-		inspectorSizingTypeString(s.Height)
-}
-
-func inspectorSizingTypeString(s sizingType) string {
-	switch s {
-	case sizingFill:
-		return "fill"
-	case sizingFixed:
-		return "fixed"
-	default:
-		return "fit"
-	}
-}
-
-func inspectorAlignString(h HorizontalAlign, v verticalAlign) string {
-	return inspectorHAlignString(h) + ", " + inspectorVAlignString(v)
-}
-
-func inspectorHAlignString(h HorizontalAlign) string {
-	switch h {
-	case HAlignEnd:
-		return "end"
-	case HAlignCenter:
-		return "center"
-	case HAlignLeft:
-		return "left"
-	case HAlignRight:
-		return "right"
-	default:
-		return "start"
-	}
-}
-
-func inspectorVAlignString(v verticalAlign) string {
-	switch v {
-	case VAlignMiddle:
-		return "middle"
-	case VAlignBottom:
-		return "bottom"
-	default:
-		return "top"
-	}
 }
