+++ gui/view_table.go
@@ -2,11 +2,8 @@ package gui
 
 import (
 	"encoding/csv"
-	"hash/fnv"
-	"log"
 	"maps"
 	"strings"
-	"sync"
 
 	"github.com/go-gui-org/go-glyph"
 )
@@ -311,365 +308,6 @@ func tableView(cfg TableCfg, w *Window) View {
 	}
 
 	return Column(outerCfg)
-}
-
-// tableBuildRow builds a single table row view.
-func tableBuildRow(
-	cfg *TableCfg, rowIdx int, columnWidths []float32,
-	cellBorder float32, selected map[int]bool,
-	multiSelect bool, colorHover Color,
-	onSelect func(map[int]bool, int, EventCtx),
-) View {
-	r := cfg.Data[rowIdx]
-	cells := make([]View, 0, len(r.Cells))
-
-	for colIdx, cell := range r.Cells {
-		cellTextStyle := cfg.TextStyle
-		if cell.TextStyle != nil {
-			cellTextStyle = *cell.TextStyle
-		} else if cell.HeadCell {
-			cellTextStyle = cfg.TextStyleHead
-		}
-
-		var colWidth float32
-		if colIdx < len(columnWidths) {
-			colWidth = columnWidths[colIdx]
-		}
-
-		hAlign := HAlignStart
-		if cell.HAlign != nil {
-			hAlign = *cell.HAlign
-		} else if cell.HeadCell {
-			hAlign = *cfg.alignHead
-		} else if colIdx < len(cfg.columnAlignments) {
-			hAlign = cfg.columnAlignments[colIdx]
-		}
-
-		var cellContent []View
-		if cell.RichText != nil {
-			cellContent = []View{
-				RTF(RTFCfg{RichText: *cell.RichText}),
-			}
-		} else if cell.Content != nil {
-			cellContent = []View{cell.Content}
-		} else {
-			cellContent = []View{
-				Text(TextCfg{
-					Text:      cell.Value,
-					TextStyle: cellTextStyle,
-				}),
-			}
-		}
-
-		cellOnClick := cell.OnClick
-		ch := colorHover
-		var cellOnHover func(EventCtx)
-		if cellOnClick != nil {
-			cellOnHover = func(ctx EventCtx) {
-				ctx.Window.SetMouseCursorPointingHand()
-				ctx.Layout.Shape.Color = ch
-			}
-		}
-
-		cells = append(cells, Column(ContainerCfg{
-			A11YRole:    AccessRoleGridCell,
-			Color:       ColorTransparent,
-			ColorBorder: cfg.ColorBorder,
-			SizeBorder:  Some(cellBorder),
-			Padding:     cfg.cellPadding,
-			Radius:      SomeF(0),
-			HAlign:      hAlign,
-			Sizing:      FixedFill,
-			Width:       colWidth,
-			OnClick:     cellOnClick,
-			OnHover:     cellOnHover,
-			Content:     cellContent,
-		}))
-	}
-
-	isSelected := selected[rowIdx]
-	rowColor := ColorTransparent
-	if isSelected {
-		rowColor = cfg.ColorSelect
-	} else if cfg.ColorRowAlt != nil && rowIdx%2 == 1 {
-		rowColor = *cfg.ColorRowAlt
-	}
-
-	rowOnClick := r.OnClick
-	ri := rowIdx
-
-	return Row(ContainerCfg{
-		Color:      rowColor,
-		Spacing:    Some(-cellBorder),
-		Padding:    NoPadding,
-		SizeBorder: NoBorder,
-		Content:    cells,
-		OnClick: func(ctx EventCtx) {
-			if rowOnClick != nil {
-				rowOnClick(ctx)
-			}
-			if onSelect != nil {
-				var newSel map[int]bool
-				if multiSelect {
-					newSel = copySelected(selected)
-				} else {
-					newSel = make(map[int]bool)
-				}
-				if newSel[ri] {
-					delete(newSel, ri)
-				} else {
-					newSel[ri] = true
-				}
-				onSelect(newSel, ri, ctx)
-			}
-		},
-		OnHover: func(ctx EventCtx) {
-			if onSelect != nil {
-				ctx.Window.SetMouseCursorPointingHand()
-				if !isSelected {
-					ctx.Layout.Shape.Color = colorHover
-				}
-			}
-		},
-	})
-}
-
-// tableFreezeLayout builds the split layout: fixed header zone
-// above a scrollable body zone.
-func tableFreezeLayout(
-	cfg *TableCfg, columnWidths []float32, cellBorder float32,
-	rowSpacing float32, selected map[int]bool,
-	multiSelect bool, colorHover Color,
-	onSelect func(map[int]bool, int, EventCtx),
-	bodyRows []View,
-	scrollID string,
-) View {
-	// Header zone: row 0 + optional separator.
-	headerViews := []View{
-		tableBuildRow(cfg, 0, columnWidths, cellBorder,
-			selected, multiSelect, colorHover, onSelect),
-	}
-
-	sepHeight := cfg.SizeBorder
-	if cfg.SizeBorderHeader > 0 {
-		sepHeight = cfg.SizeBorderHeader
-	}
-	needsSep := false
-	switch cfg.BorderStyle {
-	case TableBorderHorizontal, TableBorderHeaderOnly:
-		needsSep = true
-	}
-	if needsSep {
-		headerViews = append(headerViews, Rectangle(RectangleCfg{
-			Color:  cfg.ColorBorder,
-			Height: sepHeight,
-			Sizing: FillFixed,
-		}))
-	}
-
-	headerZone := Column(ContainerCfg{
-		Sizing:     FillFit,
-		Padding:    NoPadding,
-		Spacing:    Some(rowSpacing),
-		SizeBorder: NoBorder,
-		Content:    headerViews,
-	})
-
-	bodyCfg := ContainerCfg{
-		Sizing:     FillFill,
-		Padding:    NewPadding(0, DefaultScrollbarStyle.Size+PadXSmall, 0, 0),
-		Spacing:    Some(rowSpacing),
-		SizeBorder: NoBorder,
-		ID:         scrollID,
-		Scrollable: true,
-		Content:    bodyRows,
-		ScrollbarCfgX: &ScrollbarCfg{
-			Overflow: ScrollbarHidden,
-		},
-	}
-	if cfg.Scrollbar != ScrollbarAuto {
-		bodyCfg.ScrollbarCfgY = &ScrollbarCfg{
-			Overflow: cfg.Scrollbar,
-		}
-	}
-	bodyZone := Column(bodyCfg)
-
-	return Column(ContainerCfg{
-		ID:        cfg.ID,
-		A11YRole:  AccessRoleGrid,
-		A11YLabel: cfg.A11YLabel,
-		Color:     ColorTransparent,
-		Padding:   NoPadding,
-		Spacing:   SomeF(0),
-		Radius:    SomeF(0),
-		Sizing:    cfg.Sizing,
-		Width:     cfg.Width,
-		Height:    cfg.Height,
-		MinWidth:  cfg.MinWidth,
-		MaxWidth:  cfg.MaxWidth,
-		MinHeight: cfg.MinHeight,
-		MaxHeight: cfg.MaxHeight,
-		Content: []View{
-			headerZone,
-			bodyZone,
-		},
-	})
-}
-
-// tableColumnWidths computes column widths. When w is non-nil,
-// measures text and caches results in StateMap.
-func tableColumnWidths(cfg *TableCfg, w *Window) []float32 {
-	numCols := 0
-	for _, r := range cfg.Data {
-		if len(r.Cells) > numCols {
-			numCols = len(r.Cells)
-		}
-	}
-
-	if w == nil || w.textMeasurer == nil {
-		widths := make([]float32, numCols)
-		cw := cfg.columnWidthDefault +
-			cfg.cellPadding.Or(PaddingNone).Width()
-		for i := range widths {
-			widths[i] = cw
-		}
-		return widths
-	}
-
-	hash := tableColumnWidthHash(cfg)
-
-	if cfg.ID != "" {
-		cache := StateMap[string, tableColWidthCache](
-			w, nsTableColWidths, capModerate)
-		if cached, ok := cache.Get(cfg.ID); ok &&
-			cached.hash == hash {
-			return cached.widths
-		}
-	} else if len(cfg.Data) > 20 {
-		tableWarnNoID()
-	}
-
-	widths := tableMeasureWidths(cfg, w.textMeasurer)
-
-	if cfg.ID != "" {
-		cache := StateMap[string, tableColWidthCache](
-			w, nsTableColWidths, capModerate)
-		cache.Set(cfg.ID, tableColWidthCache{
-			hash: hash, widths: widths,
-		})
-	}
-
-	return widths
-}
-
-// tableMeasureWidths measures all columns using TextMeasurer.
-func tableMeasureWidths(
-	cfg *TableCfg, tm TextMeasurer,
-) []float32 {
-	numCols := 0
-	for _, r := range cfg.Data {
-		if len(r.Cells) > numCols {
-			numCols = len(r.Cells)
-		}
-	}
-	widths := make([]float32, numCols)
-	pad := cfg.cellPadding.Or(PaddingNone).Width()
-
-	for _, r := range cfg.Data {
-		for ci, cell := range r.Cells {
-			var tw float32
-			if cell.RichText != nil {
-				tw = tableRichTextWidth(cell.RichText, tm)
-			} else {
-				style := cfg.TextStyle
-				if cell.TextStyle != nil {
-					style = *cell.TextStyle
-				} else if cell.HeadCell {
-					style = cfg.TextStyleHead
-				}
-				tw = tm.TextWidth(cell.Value, style)
-			}
-			tw += pad
-			if tw > widths[ci] {
-				widths[ci] = tw
-			}
-		}
-	}
-
-	for i := range widths {
-		if widths[i] < cfg.columnWidthMin {
-			widths[i] = cfg.columnWidthMin
-		}
-	}
-	return widths
-}
-
-// tableRichTextWidth sums the width of each run.
-func tableRichTextWidth(rt *RichText, tm TextMeasurer) float32 {
-	var w float32
-	for _, run := range rt.Runs {
-		w += tm.TextWidth(run.Text, run.Style)
-	}
-	return w
-}
-
-// tableColumnWidthHash computes FNV-1a hash over sampled cell
-// values. Samples first, middle, and last rows.
-func tableColumnWidthHash(cfg *TableCfg) uint64 {
-	h := fnv.New64a()
-	n := len(cfg.Data)
-	_, _ = h.Write([]byte{byte(n), byte(n >> 8), byte(n >> 16), byte(n >> 24)})
-	indices := make([]int, 0, 3)
-	if n > 0 {
-		indices = append(indices, 0)
-	}
-	if n > 2 {
-		indices = append(indices, n/2)
-	}
-	if n > 1 {
-		indices = append(indices, n-1)
-	}
-	for _, i := range indices {
-		for _, cell := range cfg.Data[i].Cells {
-			_, _ = h.Write([]byte(cell.Value))
-		}
-	}
-	return h.Sum64()
-}
-
-var tableWarnNoID = sync.OnceFunc(func() {
-	log.Printf("gui.Table: table with >20 rows has no ID; " +
-		"column width caching disabled")
-})
-
-// tableEstimateRowHeight estimates row height from TextStyle,
-// cell padding, and border.
-func tableEstimateRowHeight(cfg *TableCfg, w *Window) float32 {
-	style := cfg.TextStyle
-	height := style.Size
-	if w != nil && w.textMeasurer != nil {
-		height = w.textMeasurer.FontHeight(style)
-	}
-	return height + cfg.cellPadding.Or(PaddingNone).Height()
-}
-
-// ClearTableCache removes cached column widths for the given
-// table ID.
-func (w *Window) clearTableCache(id string) {
-	cache := StateMapRead[string, tableColWidthCache](
-		w, nsTableColWidths)
-	if cache != nil {
-		cache.Delete(id)
-	}
-}
-
-// ClearAllTableCaches removes all cached table column widths.
-func (w *Window) clearAllTableCaches() {
-	cache := StateMapRead[string, tableColWidthCache](
-		w, nsTableColWidths)
-	if cache != nil {
-		cache.Clear()
-	}
 }
 
 // TR creates a table row from the given cells.
